SQUASH: Come back to this

+++ qiime2/core/cache.py
@@ -228,10 +228,11 @@ def _exit_cleanup():
         # does not create a process pool (on Mac this atexit is invoked on
         # workers). It could also happen if someone deleted the process pool
         # but... They probably shouldn't do that
-        if os.path.exists(target):
+        if os.path.exists(cache.path):
             with cache.lock:
-                shutil.rmtree(target)
-                cache.garbage_collection()
+                if os.path.exists(target):
+                    shutil.rmtree(target)
+                    cache.garbage_collection()
 
 
 def monitor_thread(cache_dir, is_done):
